[af-146] style: restful 한 구조를 위해 store 제거
- route 에 store 제거
- 패키지 구조에 store 제거
- class name에 store 제거
