+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/product/dto/ProductRequest.java
@@ -1,0 +1,77 @@
+package com.drunkenlion.alcoholfriday.domain.admin.product.dto;
+
+import com.drunkenlion.alcoholfriday.domain.category.entity.Category;
+import com.drunkenlion.alcoholfriday.domain.maker.entity.Maker;
+import com.drunkenlion.alcoholfriday.domain.product.entity.Product;
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.*;
+
+import java.math.BigDecimal;
+
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Schema(description = "제품 입력 요청 항목")
+public class ProductRequest {
+    @Schema(description = "카테고리 소분류 고유 아이디")
+    private Long categoryLastId;
+
+    @Schema(description = "제품명")
+    private String name;
+
+    @Schema(description = "제조사 고유 아이디")
+    private Long makerId;
+
+    @Schema(description = "제품 원가")
+    private BigDecimal price;
+
+    @Schema(description = "재고 수량")
+    private Long quantity;
+
+    @Schema(description = "술 도수")
+    private Long alcohol;
+
+    @Schema(description = "제품 재료")
+    private String ingredient;
+
+    @Schema(description = "술 단맛")
+    private Long sweet;
+
+    @Schema(description = "술 신맛")
+    private Long sour;
+
+    @Schema(description = "술 청량감")
+    private Long cool;
+
+    @Schema(description = "술 바디감")
+    private Long body;
+
+    @Schema(description = "술 밸런스")
+    private Long balence;
+
+    @Schema(description = "술 향기")
+    private Long insense;
+
+    @Schema(description = "술 목넘김")
+    private Long throat;
+
+    public static Product toEntity(ProductRequest request, Category category, Maker maker) {
+        return Product.builder()
+                .name(request.getName())
+                .price(request.getPrice())
+                .quantity(request.getQuantity())
+                .alcohol(request.getAlcohol())
+                .ingredient(request.getIngredient())
+                .sweet(request.getSweet())
+                .sour(request.getSour())
+                .cool(request.getCool())
+                .body(request.getBody())
+                .balence(request.getBalence())
+                .insense(request.getInsense())
+                .throat(request.getThroat())
+                .category(category)
+                .maker(maker)
+                .build();
+    }
+}
